Fixed some bugs, mostly around the FPU.  Now Linux can start without the "cp memory exhausted" warning.  And Age of Empires works correctly now.
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@744 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/cpu/core_normal/Prefix_none.java
@@ -1157,7 +1157,7 @@ public class Prefix_none extends StringOp {
                 /*Bit8u*/int rm=Fetchb();/*PhysPt*/int eaa=getEaa(rm);
                 bound_min=(short)Memory.mem_readw(eaa);
                 bound_max=(short)Memory.mem_readw(eaa+2);
-                if ( ((Modrm.Getrw[rm].word()) < bound_min) || ((Modrm.Getrw[rm].word()) > bound_max) ) {
+                if ( (((short)Modrm.Getrw[rm].word()) < bound_min) || (((short)Modrm.Getrw[rm].word()) > bound_max) ) {
                     return EXCEPTION(5);
                 }
                 return HANDLED;
